Remove legacy useTokenDataWebSocket.ts and centralize TokenData types
